♻️ Del opp FullførKjørelisteFane i hooks for lesbarhet
Trekker ut to dedikerte hooks:
- useKjørelisteBrev: henter/genererer brev ved mount
- useFullførKjøreliste: håndterer fullføring og validering

FullførKjørelisteFane er nå ren layout/render.
